# Conflicts:
#	capacity_parent/capacity_common/src/main/java/com/zhidejiaoyu/common/mapper/StudentEquipmentMapper.java
#	capacity_parent/capacity_common/src/main/java/com/zhidejiaoyu/common/mapper/StudentEquipmentMapper.xml

+++ capacity_parent/capacity_front/src/main/java/com/zhidejiaoyu/student/business/controller/PersonalCentreController.java
@@ -269,7 +269,7 @@ public class PersonalCentreController extends BaseController {
     }
 
     @GetMapping("/getLucky")
-    public Object getLucky(Integer studentId, HttpSession session) {
+    public Object getLucky(Long studentId, HttpSession session) {
         return personalCentreService.getLucky(studentId, session);
     }
